api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.1
openai>=1.55.0
azure-storage-blob>=12.19.0
azure-cognitiveservices-speech>=1.35.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.27.0
